src/policy: move all policy related code into `policy` mod
