+++ src/lib.rs
@@ -1,11 +1,15 @@
-mod builder;
 mod complex;
-mod ext;
-mod policy;
-pub mod preset;
+pub mod policy;
 
-pub use builder::PolicyBuilder;
-pub use ext::SanitizeExt;
-pub use policy::{AllowAllPolicy, DenyAllPolicy, PermissivePolicy, RestrictivePolicy};
-pub use policy::{AttributeRule, Policy};
-pub use policy::{Permissive, Restrictive};
+pub use policy::*;
+
+
+/// A base sanitization directive, which allows all elements and attributes,
+/// excluding listed in policy.
+#[derive(Debug, Clone, Copy)]
+pub struct Permissive;
+
+/// A base sanitization directive, which restricts all elements and attributes,
+/// excluding listed in policy.
+#[derive(Debug, Clone, Copy)]
+pub struct Restrictive;
